kits/option_b/requirements.txt
langchain
langgraph
langchain-openai
langchain-anthropic
python-dotenv
httpx
pydantic
